時間マネージャーfloat返す関数にした

+++ 3D_Base/Source/GameObject/MeshObject/StaticMesh/Ground/CGround.cpp
@@ -70,7 +70,7 @@ void CGround::Draw(
 void CGround::ChangeColorOfGround()
 {
 	//経過時間を取得.
-	float t = static_cast<float>(CTimeManager::GetTotalTime());
+	float t = CTimeManager::GetTotalTime();
 
 	//全体の時間の現在の割合.
 	float progress =
@@ -94,7 +94,7 @@ void CGround::ChangeColorOfGround()
 void CGround::ShakeGround()
 {
 	//経過時間を取得.
-	float t = static_cast<float>(CTimeManager::GetTotalTime());
+	float t = CTimeManager::GetTotalTime();
 
 	//揺れている時間の現在の割合.
 	float progress = 
@@ -109,7 +109,7 @@ void CGround::ShakeGround()
 void CGround::FallDownGround()
 {
 	//デルタタイムを取得.
-	float dt = static_cast<float>(CTimeManager::GetDeltaTime());
+	float dt = CTimeManager::GetDeltaTime();
 
 	m_Velocity += dt * -m_Gravity;
 	m_vPosition.y -= dt * m_Velocity * m_FallSpeed;	//地面を下へ.
