+++ 3D_Base/Source/GameObject/MeshObject/StaticMesh/PlayerBase/PlayerState/PlayerActionState/PlayerGetUpState/CPlayerGetUpState.cpp
@@ -57,7 +57,7 @@ void CPlayerGetUpState::Exit()
 
 void CPlayerGetUpState::Update()
 {
-	float t = static_cast<float>(CTimeManager::GetTotalTime());
+	float t = CTimeManager::GetTotalTime();
 
 	//現在の経過時間と開始時間の差が終了時間を上回ったら.
 	if (t - m_StartTime > m_EndTime)
