+++ 3D_Base/Source/GameObject/MeshObject/StaticMesh/PlayerBase/PlayerState/PlayerActionState/PlayerPickupState/CPlayerPickupState.cpp
@@ -68,7 +68,7 @@ void CPlayerPickupState::Exit()
 void CPlayerPickupState::Update()
 {
 	//ゲーム全体の経過時間.
-	float t = static_cast<float>(CTimeManager::GetTotalTime());
+	float t = CTimeManager::GetTotalTime();
 
 	//現在の経過時間と開始時間の差が終了時間を上回ったら.
 	if (t - m_StartTime > m_EndTime)
